Clean up the package
Move test to the top level directory

Move reqs to a dedicated directory

Update setup.py

Remove stale stuff

Drop distutils and distribute
