--- .hgignore
@@ -1,13 +1,0 @@
-syntax: glob
-*.pyc
-.project
-.pydevproject
-build
-env
-data
-dist
-.settings
-.coverage
-*.egg*
-*.log
-coilmq.cfg
