+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = --doctest-modules --ignore=setup.py --cov=coilmq
+addopts = --doctest-modules --ignore=setup.py --cov=coilmq --cov=tests
