+++ requirements/build.txt
@@ -1,4 +1,5 @@
 SQLAlchemy
 fakeredis
+mock
 pytest
 pytest-cov
