+++ requirements/prod.txt
@@ -1,6 +1,5 @@
 python-daemon
 pid
-py==1.4.30
-wheel==0.24.0
+wheel
 six
 click
