+++ setup.py
@@ -2,19 +2,9 @@
 Simple, lightweight, and easily extensible STOMP message broker.
 """
 import os.path
-import warnings
-import re
 
-
-try:
-    import setuptools
-except ImportError:
-    from distribute_setup import use_setuptools
-    use_setuptools()
 from setuptools import setup, find_packages
 from setuptools.command.test import test
-
-from distutils.core import setup, Command
 
 
 class PyTest(test):
@@ -31,17 +21,6 @@ class PyTest(test):
 
 version = '0.6'
 
-news = os.path.join(os.path.dirname(__file__), 'docs', 'news.txt')
-news = open(news).read()
-parts = re.split(r'([0-9\.]+)\s*\n\r?-+\n\r?', news)
-found_news = ''
-for i in range(len(parts)-1):
-    if parts[i] == version:
-        found_news = parts[i+i]
-        break
-if not found_news:
-    warnings.warn('No news for this version found.')
-
 long_description = """
 The provided server implementation for CoilMQ uses the Python SocketServer
 libraries; however, CoilMQ is only loosely coupled to this server
@@ -49,19 +28,22 @@ implementation.  It could be used with other socket implementations.
 
 The CoilMQ core classes and bundled storage implementations are built to be
 thread-safe.
+
+
 """
 
-if found_news:
-    title = 'Changes in %s' % version
-    long_description += "\n%s\n%s\n" % (title, '-'*len(title))
-    long_description += found_news
+
+def read(fname):
+    with open(os.path.join(os.path.dirname(__file__), fname)) as f:
+        return f.read()
+
 
 setup(
     name='CoilMQ',
     version=version,
     description=__doc__,
-    long_description=long_description,
-    keywords='stomp server broker',
+    long_description=long_description + read('docs/news.txt'),
+    keywords='stomp server broker messaging',
     license='Apache',
     author='Hans Lellelid',
     author_email='hans@xmpl.org',
@@ -69,23 +51,22 @@ setup(
     packages=find_packages(exclude=['ez_setup', 'distribute_setup', 'tests', 'tests.*']),
     package_dir={'coilmq':  'coilmq'},
     package_data={'coilmq': ['config/*.cfg*', 'tests/resources/*']},
-    zip_safe=False, # We use resource_filename for logging configuration and some unit tests.
+    zip_safe=False,  # We use resource_filename for logging configuration and some unit tests.
     include_package_data=True,
-    tests_require=['pytest', 'pytest-cov'],
-    cmdclass = {'test': PyTest},
-    install_requires=[
-          'distribute',
-    ],
+    tests_require=read('requirements/build.txt').splitlines(),
+    cmdclass={'test': PyTest},
+    install_requires=read('requirements/prod.txt').splitlines(),
     extras_require={
-        'daemon': ['python-daemon'],
-        'sqlalchemy': ['SQLAlchemy']
+        'sqlalchemy': ['SQLAlchemy'],
+        'redis': ['redis']
     },
     classifiers=[
         "Development Status :: 4 - Beta",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python :: 2.6",
         "Programming Language :: Python :: 2.7",
+        "Programming Language :: Python :: 3.4",
+        "Programming Language :: Python :: 3.5",
         "Operating System :: OS Independent",
         "Topic :: Software Development :: Libraries :: Python Modules",
         "Topic :: Communications",
