+++ tests/functional/test_basic.py
@@ -9,7 +9,7 @@ import six
 
 from coilmq.auth.simple import SimpleAuthenticator
 from coilmq.util import frames
-from coilmq.tests.functional import BaseFunctionalTestCase, Empty as QueueEmpty
+from tests.functional import BaseFunctionalTestCase, Empty as QueueEmpty
 
 __authors__ = ['"Hans Lellelid" <hans@xmpl.org>']
 __copyright__ = "Copyright 2009 Hans Lellelid"
