+++ tests/resources/auth.ini
