+++ tests/store/test_dbm.py
@@ -1,16 +1,15 @@
 """
 Test DBM queue storage.
 """
-import unittest
-import tempfile
 import shutil
-import uuid
+import tempfile
 import time
+import unittest
+import uuid
 
-from coilmq.util.frames import Frame
 from coilmq.store.dbm import DbmQueue
-
-from coilmq.tests.store import CommonQueueTest
+from coilmq.util.frames import Frame
+from tests.store import CommonQueueTest
 
 __authors__ = ['"Hans Lellelid" <hans@xmpl.org>']
 __copyright__ = "Copyright 2009 Hans Lellelid"
