+++ tests/test_buffer.py
