+++ tests/test_engine.py
@@ -4,10 +4,8 @@ Tests for the transport-agnostic engine module.
 import unittest
 
 from coilmq.engine import StompEngine
-
-from coilmq.tests.mock import (MockAuthenticator, MockConnection, MockQueueManager, MockTopicManager)
 from coilmq.util.frames import Frame, ReceiptFrame
-from coilmq.util import frames
+from tests.mock import (MockAuthenticator, MockConnection, MockQueueManager, MockTopicManager)
 
 __authors__ = ['"Hans Lellelid" <hans@xmpl.org>']
 __copyright__ = "Copyright 2009 Hans Lellelid"
